+++ packages/react-aria-components/scripts/create-react-aria-components/helpers/createProjectFolder.js
@@ -1,0 +1,8 @@
+import fs from 'fs';
+
+export const createProjectFolder = (projectName) => {
+  if (!fs.existsSync(projectName)) {
+    fs.mkdirSync(projectName);
+  }
+    // TODO handle error
+};
