+++ packages/react-aria-components/scripts/create-react-aria-components/package.json
@@ -4,6 +4,9 @@
   "description": "A command line tool for creating React Aria Components from templates. Easily start a new component library, or add components to an existing project. ",
   "main": "index.js",
   "type": "module",
+  "bin": {
+    "create-react-aria-components": "./index.js"
+  },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -15,7 +18,7 @@
   "license": "Apache-2.0",
   "dependencies": {
     "chalk": "^5.2.0",
-    "prompts": "^2.4.2",
-    "react-aria-components": "1.0.0-alpha.3"
+    "got": "^12.6.0",
+    "prompts": "^2.4.2"
   }
 }
