+++ packages/react-aria-components/scripts/create-react-aria-components/templates/vanilla-css/components/Button/Button.mdx
@@ -1,4 +1,4 @@
-import {Button} from './Button''
+import {Button} from './Button'
 
 <Meta title="Button" component={Button} />
