+++ packages/react-aria-components/scripts/create-react-aria-components/templates/vanilla-css/components/Button/index.ts
