+++ packages/react-aria-components/scripts/create-react-aria-components/templates/vanilla-css/index.ts
@@ -1,1 +1,1 @@
-export {Button} from './Button';
+export {Button} from './components/Button';
